+++ woo-payzen-payment/class-wc-gateway-payzensubscription.php
@@ -77,6 +77,8 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
         // Order needs payment filter.
         add_filter('woocommerce_order_needs_payment', array($this, 'payzen_order_needs_payment'), 10, 2);
 
+        add_action('manage_shop_subscription_posts_custom_column', array($this, 'payzen_display_subscription_error_msg'), 2);
+
         $this->subscriptions_handler->init_hooks();
     }
 
@@ -119,8 +121,17 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
     {
         global $woocommerce;
 
-        $cust_id = $this->get_customer_property($woocommerce->customer, 'id');
-        if (! $cust_id) {
+        // In case of changing payment method of an existing subscription.
+        // At this stage all conditions of is_available_for_subscriptions are guaranteed so we return true.
+        if ($this->subscriptions_handler && $this->subscriptions_handler->is_subscription_update()) {
+            return true;
+        }
+
+        $cust_id = self::get_customer_property($woocommerce->customer, 'id');
+
+        // Allow subscription when no client is connected and "Allow customers to create an account during checkout" is enabled.
+        if (! $cust_id && (get_option('woocommerce_enable_signup_and_login_from_checkout') !== 'yes')
+            && (get_option('woocommerce_enable_signup_from_checkout_for_subscriptions') !== 'yes')) {
             return false;
         }
 
@@ -133,11 +144,23 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
         return true;
     }
 
+    public function payment_fields()
+    {
+        parent::payment_fields();
+
+        if ($this->subscriptions_handler && $this->subscriptions_handler->is_subscription_update()) {
+            $order_id = get_query_var('order-pay');
+            $order = new WC_Order((int) $order_id);
+            $method = self::get_order_property($order, 'payment_method');
+            echo '<input type="hidden" id="payzensubscription_old_pm" name="payzensubscription_old_pm" value="' . $method . '">';
+        }
+    }
+
     protected function payment_by_alias_view($html, $force_redir = true)
     {
         global $woocommerce;
 
-        $cust_id = $this->get_customer_property($woocommerce->customer, 'id');
+        $cust_id = self::get_customer_property($woocommerce->customer, 'id');
         $saved_subsc_masked_pan = get_user_meta((int) $cust_id, $this->id . '_masked_pan', true);
 
         // Recover card brand if saved with masked pan and check if logo exists.
@@ -192,21 +215,19 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
     {
         parent::payzen_fill_request($order);
 
+        $cust_id = self::get_order_property($order, 'user_id');
+        $saved_identifier = $this->get_cust_identifier($cust_id);
+        $is_identifier_active = $this->is_cust_identifier_active($cust_id);
+
         $info = $this->subscriptions_handler->subscription_info($order);
 
         if (is_array($info) && ! empty($info)) {
             $currency = PayzenApi::findCurrencyByAlphaCode(get_woocommerce_currency());
-            $cust_id = $this->get_order_property($order, 'user_id');
 
             $this->payzen_request->set('sub_amount', $currency->convertAmountToInteger($info['amount']));
             $this->payzen_request->set('sub_currency', $currency->getNum()); // Same as general order currency.
 
-            $desc = 'RRULE:FREQ=' . $info['frequency']. ';INTERVAL=' . $info['interval'];
-            if (isset($info['end_date']) && $info['end_date']) {
-                $desc .= ';UNTIL=' . $info['end_date'];
-            }
-
-            $this->payzen_request->set('sub_desc', $desc);
+            $this->payzen_request->set('sub_desc', $this->get_rrule($info));
 
             $this->payzen_request->set('sub_effect_date', $info['effect_date']);
 
@@ -216,12 +237,12 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
                 $this->payzen_request->set('sub_init_amount_number', $info['init_number']);
             }
 
-            if ($order->get_total() > 0) {
+            $order_amount = $order->get_total();
+            if ($order_amount > 0) {
                 $this->payzen_request->set('page_action', 'REGISTER_PAY_SUBSCRIBE');
+                $this->payzen_request->set('identifier', null);
             } else {
                 // Only subscriptions.
-                $saved_identifier = $this->get_cust_identifier($cust_id);
-                $is_identifier_active = $this->is_cust_identifier_active($cust_id);
                 if ($saved_identifier && $is_identifier_active) {
                     $this->payzen_request->set('identifier', $saved_identifier);
                     $this->payzen_request->set('page_action', 'SUBSCRIBE');
@@ -229,6 +250,19 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
                     $this->payzen_request->set('page_action', 'REGISTER_SUBSCRIBE');
                 }
             }
+        } elseif ($saved_identifier) {
+            // Called from change payment action.
+            $this->payzen_request->set('amount', 0);
+            $this->payzen_request->set('identifier', $saved_identifier);
+            $this->payzen_request->set('page_action', 'REGISTER_UPDATE');
+        }
+
+        // $order_id is an id of a subscription.
+        $order_id = self::get_order_property($order, 'id');
+        if ($parent_order = $this->subscriptions_handler->get_parent_order($order_id)) {
+            $this->payzen_request->set('order_id', self::get_order_property($parent_order, 'id'));
+            $this->payzen_request->addExtInfo('order_key', self::get_order_property($parent_order, 'order_key'));
+            $this->payzen_request->addExtInfo('subsc_id', $order_id);
         }
     }
 
@@ -236,6 +270,7 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
     {
         $key = $this->testmode ? $this->get_general_option('test_private_key') : $this->get_general_option('prod_private_key');
         if (! $key) {
+            set_transient('payzen_cancelled_subscription_error_msg', sprintf(__('Subscription is cancelled only in WooCommerce. Please, consider cancelling the subscription in %s Back Office.', 'woo-payzen-payment'), 'PayZen'));
             $this->log("Subscription #{$subscription_id} cannot be cancelled on gateway for order #$order_id: private key is not configured.");
             return;
         }
@@ -244,7 +279,7 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
 
         $order = new WC_Order((int) $order_id);
 
-        $cust_id = $this->get_order_property($order, 'user_id');
+        $cust_id = self::get_order_property($order, 'user_id');
         $saved_identifier = $this->get_cust_identifier($cust_id);
 
         $params = array(
@@ -272,25 +307,22 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
                 $this->cancel_transaction($transaction);
             }
         } catch (Exception $e) {
-            $this->log("Subscription cancel exception for order #$order_id with code {$e->getCode()}: {$e->getMessage()}.");
+            set_transient('payzen_cancelled_subscription_error_msg', sprintf(__('An error has occurred during the cancellation or update of the subscription. Please consult the %s logs for more details.', 'woo-payzen-payment'), 'PayZen'));
+            $this->log("Subscription cancel exception for order #$order_id with code {$e->getCode()}: {$e->getMessage()}");
         }
     }
 
     private function cancel_transaction($transaction)
     {
-        if ($transaction['status'] !== 'RUNNING' ){
+        if ($transaction['status'] !== 'RUNNING' ) {
             return;
         }
 
-        $order_id =  $transaction['orderDetails']['orderId'];
+        $order_id = $transaction['orderDetails']['orderId'];
 
         $this->log("Cancelling transaction with UUID #{$transaction['uuid']} for order #$order_id.");
 
         $key = $this->testmode ? $this->get_general_option('test_private_key') : $this->get_general_option('prod_private_key');
-        if (! $key) {
-            $this->log("Transaction with UUID #{$transaction['uuid']} cannot be cancelled on gateway for order #$order_id: private key is not configured.");
-            return;
-        }
 
         try {
             $params = array(
@@ -307,6 +339,7 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
             $result = $client->post('V4/Transaction/Cancel', json_encode($params));
             PayzenRestTools::checkResult($result, 'CANCELLED');
         } catch (Exception $e) {
+            set_transient('payzen_cancelled_subscription_error_msg', sprintf(__('An error has occurred during the cancellation or update of the subscription. Please consult the %s logs for more details.', 'woo-payzen-payment'), 'PayZen'));
             $this->log("Transaction cancel exception for order #$order_id with code {$e->getCode()}: {$e->getMessage()}.");
         }
     }
@@ -315,6 +348,7 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
     {
         $key = $this->testmode ? $this->get_general_option('test_private_key') : $this->get_general_option('prod_private_key');
         if (! $key) {
+            set_transient('payzen_update_subscription_error_msg', sprintf(__('Subscription is updated only in WooCommerce. Please, consider making necessary changes in %s Back Office.', 'woo-payzen-payment'), 'PayZen'));
             $this->log("Subscription #{$subscription_id} for order #$order_id cannot be updated on gateway: private key is not configured.");
             return;
         }
@@ -323,12 +357,13 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
 
         $order = new WC_Order((int) $order_id);
 
-        $cust_id = $this->get_order_property($order, 'user_id');
+        $cust_id = self::get_order_property($order, 'user_id');
         $saved_identifier = $this->get_cust_identifier($cust_id);
 
         // Re-generate subscriptions data from updated order.
         $info = $this->subscriptions_handler->subscription_info($order);
         if (! is_array($info) || empty($info)) {
+            set_transient('payzen_update_subscription_error_msg', sprintf(__('An error has occurred during the cancellation or update of the subscription. Please consult the %s logs for more details.', 'woo-payzen-payment'), 'PayZen'));
             $this->log("Empty subscription info returned for order #$order_id. Cannot update subscription #{$subscription_id}.");
             return;
         }
@@ -336,14 +371,9 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
         // Get subscription effect date in ISO8601 format.
         $date_time = strtotime($info['effect_date']);
         if ($date_time < time()) {
+            set_transient('payzen_update_subscription_error_msg', sprintf(__('An error has occurred during the cancellation or update of the subscription. Please consult the %s logs for more details.', 'woo-payzen-payment'), 'PayZen'));
             $this->log("Cannot update subscription #{$subscription_id} for order #$order_id. Effect date has passed.");
             return;
-        }
-
-        // Compute new rrule.
-        $desc = 'RRULE:FREQ=' . $info['frequency']. ';INTERVAL=' . $info['interval'];
-        if (isset($info['end_date']) && $info['end_date']) {
-            $desc .= ';UNTIL=' . $info['end_date'];
         }
 
         $currency = PayzenApi::findCurrencyByAlphaCode(get_woocommerce_currency());
@@ -354,7 +384,7 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
             'amount' => $currency->convertAmountToInteger($info['amount']),
             'currency' => $currency->getAlpha3(),
             'effectDate' => date(DateTime::ATOM, $date_time),
-            'rrule' => $desc
+            'rrule' => $this->get_rrule($info)
         );
 
         // Initial amount.
@@ -376,6 +406,7 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
             //Subscription cancelled successfully.
             $this->log("Subscription #{$subscription_id} updated successfully for order #$order_id.");
         } catch (Exception $e) {
+            set_transient('payzen_update_subscription_error_msg', sprintf(__('An error has occurred during the cancellation or update of the subscription. Please consult the %s logs for more details.', 'woo-payzen-payment'), 'PayZen'));
             $this->log("Subscription update exception for order #$order_id with code {$e->getCode()}: {$e->getMessage()}");
         }
     }
@@ -394,31 +425,98 @@ class WC_Gateway_PayzenSubscription extends WC_Gateway_PayzenStd
             'operationType' => 'DEBIT'
         );
 
-        $get_order_response = $client->post('V4/Order/Get', json_encode($params));
-        PayzenRestTools::checkResult($get_order_response);
+        try {
+            $get_order_response = $client->post('V4/Order/Get', json_encode($params));
+            PayzenRestTools::checkResult($get_order_response);
 
-        // Order transactions organized by sequence numbers.
-        $trans_by_sequence = array();
-        foreach ($get_order_response['answer']['transactions'] as $transaction) {
-            $sequence_number = $transaction['transactionDetails']['sequenceNumber'];
-            // Unpaid transactions are not considered.
-            if ($transaction['status'] !== 'UNPAID') {
-                $trans_by_sequence[$sequence_number] = $transaction;
+            // Order transactions organized by sequence numbers.
+            $trans_by_sequence = array();
+            foreach ($get_order_response['answer']['transactions'] as $transaction) {
+                $sequence_number = $transaction['transactionDetails']['sequenceNumber'];
+                // Unpaid transactions are not considered.
+                if ($transaction['status'] !== 'UNPAID') {
+                    $trans_by_sequence[$sequence_number] = $transaction;
+                }
             }
-        }
 
-        ksort($trans_by_sequence);
-        return array_reverse($trans_by_sequence);
+            ksort($trans_by_sequence);
+            return array_reverse($trans_by_sequence);
+        } catch (Exception $e) {
+            set_transient('payzen_cancelled_subscription_error_msg', sprintf(__('An error has occurred during the cancellation or update of the subscription. Please consult the %s logs for more details.', 'woo-payzen-payment'), 'PayZen'));
+            $this->log("Order transactions processing exception for order #$order_id with code {$e->getCode()}: {$e->getMessage()}.");
+        }
     }
 
     public function payzen_order_needs_payment($is_active, $order)
     {
         global $woocommerce;
 
-        if (($order->get_total() == 0) && ($this->get_order_property($order, 'payment_method') === $this->id)) {
+        if (($order->get_total() == 0) && (self::get_order_property($order, 'payment_method') === $this->id)) {
             return $this->subscriptions_handler && $this->subscriptions_handler->cart_contains_subscription($woocommerce->cart);
         }
 
         return $is_active;
+    }
+
+    public function payzen_display_subscription_error_msg($column)
+    {
+        $payzen_cancelled_subscription_error_msg = get_transient('payzen_cancelled_subscription_error_msg');
+        if ($payzen_cancelled_subscription_error_msg) {
+            delete_transient('payzen_cancelled_subscription_error_msg');
+            ?>
+            <script type="text/javascript">
+                jQuery(document).ready(function() {
+                    if (! jQuery('#payzen_cancelled_subscription_error').length) {
+                        jQuery('a.page-title-action').after('<div id="payzen_cancelled_subscription_error" class="error notice is-dismissible"><p><?php echo addslashes($payzen_cancelled_subscription_error_msg); ?></p><button type="button" class="notice-dismiss" onclick="this.parentElement.remove()"><span class="screen-reader-text"><?php echo esc_html__( 'Dismiss this notice.', 'woocommerce' )  ?></span></button></div>');
+                    }
+                });
+            </script>
+            <?php
+        }
+    }
+
+    /**
+     * Process the payment and return the result.
+     **/
+    public function process_payment($order_id)
+    {
+        global $woocommerce;
+
+        if (isset($_POST['payzensubscription_old_pm'])) {
+            set_transient($this->id . '_change_payment_' . $order_id, $_POST['payzensubscription_old_pm']);
+        }
+
+        $order = new WC_Order($order_id);
+
+        if (version_compare($woocommerce->version, '2.1.0', '<')) {
+            $pay_url = add_query_arg('order', self::get_order_property($order, 'id'), add_query_arg('key', self::get_order_property($order, 'order_key'), get_permalink(woocommerce_get_page_id('pay'))));
+        } else {
+            $pay_url = $order->get_checkout_payment_url(true);
+        }
+
+        return array(
+            'result' => 'success',
+            'redirect' => $pay_url
+        );
+    }
+
+    private function get_rrule($subscription_info) {
+        $desc = 'RRULE:FREQ=' . $subscription_info['frequency'] . ';INTERVAL=' . $subscription_info['interval'];
+
+        $date_time = strtotime($subscription_info['effect_date']);
+        $array_date = getdate($date_time);
+        if (in_array($subscription_info['frequency'], array('MONTHLY', 'YEARLY')) && ((int) $array_date['mday'] > 28)) {
+            $desc .= ';BYMONTHDAY=28,29,30,31;BYSETPOS=-1';
+
+            if ($subscription_info['frequency'] === 'YEARLY') {
+                $desc .= ';BYMONTH=' . $array_date['mon'];
+            }
+        }
+
+        if (isset($subscription_info['end_date']) && $subscription_info['end_date']) {
+            $desc .= ';UNTIL=' . $subscription_info['end_date'];
+        }
+
+        return $desc;
     }
 }