+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+= 1.9.1, 2021-06-21=
+* Bug fix: Do not create two transactions when trial is disabled for a subscription.
+* Bug fix: Fatal error when modifying payment for a subscription in My account > subscriptions.
+* Bug fix: Propose dynamically added payment means in "Other payment means" section.
+* Bug fix: Propose subscription payment method when client account creation during checkout is enabled.
+* Bug fix: Adjust rrule for dates at the end of the month when creating subscriptions.
+* Manage retrocompatibility with already validated orders (do not check order key) when processing subscriptions.
+* Manage subscription creation from gateway Back Office.
+* Improve error management on subscription actions (cancel and update).
+* Send the relevant part of the current PHP version in vads_contrib field.
+* Improve support e-mails display.
+
 = 1.9.0, 2021-04-21=
 * [subscr] Manage subscriptions with WooCommerce Subscriptions (including subscription update and cancellation).
 * [franfinance] Added new Franfinance submodule.